+++ src/vue/modules/update-asset-form-simplified/mixins/manage-asset-request.mixin.js
@@ -1,11 +1,9 @@
-import { base } from '@tokend/js-sdk'
+import { base, Document } from '@tokend/js-sdk'
 import { REQUEST_STATES } from '@/js/const/request-states.const'
 
 import { api } from '@/api'
-import { uploadDocuments } from '@/js/helpers/upload-documents'
 
 import { UpdateAssetRequest } from '../wrappers/update-asset-request'
-import { DocumentContainer } from '@/js/helpers/DocumentContainer'
 import { mapGetters } from 'vuex'
 import { vuexTypes } from '@/vuex/index'
 import { DateUtil } from '@/js/utils'
@@ -72,11 +70,7 @@ export default {
     },
 
     async submitUpdateAssetRequest (requestId) {
-      const assetDocuments = [
-        this.collectedAttributes.logo,
-        this.collectedAttributes.terms,
-      ]
-      await uploadDocuments(assetDocuments)
+      await Document.uploadDocumentsDeep(this.collectedAttributes)
 
       await api.postOperations(
         this.$buildAssetUpdateRequestOperation(requestId),
@@ -84,17 +78,13 @@ export default {
     },
 
     $buildAssetUpdateRequestOperation (requestId) {
-      const logo = this.collectedAttributes.logo
-      const terms = this.collectedAttributes.terms
-
       const opts = {
         requestID: requestId ? String(requestId) : NEW_UPDATE_ASSET_REQUEST_ID,
         code: this.collectedAttributes.code,
         policies: this.collectedAttributes.policies,
         creatorDetails: {
           name: this.collectedAttributes.name,
-          logo: this.$getDocumentDetailsOrEmptyDocument(logo),
-          terms: this.$getDocumentDetailsOrEmptyDocument(terms),
+          logo: this.collectedAttributes.logo,
           description: this.collectedAttributes.description,
           stellar: {},
           ...(this.collectedAttributes.expirationDate
@@ -109,14 +99,6 @@ export default {
 
     collectAssetAttributes (newAttributes) {
       Object.assign(this.collectedAttributes, newAttributes)
-    },
-
-    $getDocumentDetailsOrEmptyDocument (doc) {
-      if (doc instanceof DocumentContainer) {
-        return doc.getDetailsForSave()
-      }
-
-      return DocumentContainer.getEmptyDetailsForSave()
     },
   },
 }