+++ src/vue/pages/Settings.vue
@@ -26,8 +26,6 @@ import { vueRoutes } from '@/vue-router/routes'
 import { vuexTypes } from '@/vuex'
 import { mapGetters } from 'vuex'
 
-import { REQUEST_STATES_STR } from '@/js/const/request-states.const'
-
 export default {
   name: 'settings',
   components: {
@@ -37,11 +35,11 @@ export default {
     vueRoutes,
   }),
   computed: {
-    ...mapGetters({
-      kycState: vuexTypes.kycState,
-    }),
+    ...mapGetters([
+      vuexTypes.kycRequest,
+    ]),
     verificationTabId () {
-      if (this.kycState === REQUEST_STATES_STR.approved) {
+      if (this.kycRequest.isApproved) {
         return 'settings-page.my-account-tab'
       } else {
         return 'settings-page.verification-tab'